final
Create a Reaction Time Monitor (RTM) that can indicate how quickly an user can respond to a
stimulus. In operation, the RTM is initialized when a “start” button is pressed. Immediately after start,
the 7seg display shows all 0’s. A random time later (between about 1 and 10 seconds), a “react now”
LED illuminates and a millisecond timer starts. When the timer starts, the 7seg display shows the
timer value as it counts up in milliseconds. After the LED turns on, the user presses a “react” button
as soon as possible to stop the timer. The stopped timer will contain the number of milliseconds
between the react LED turning on and the button being pressed, and that time will be shown on the
7seg display. Pressing the start button again will clear the timer and begin a new round.